fix: redesign like increasing system

+++ src/main/java/com/bondarenko/apps/boot_js_app/domain/json/JWT.java
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class JWT extends Author {
     @JsonProperty("exp")
-    private int exp;
+    private long exp;
     @JsonProperty("iat")
-    private int iat;
+    private long iat;
     @JsonProperty("iss")
     private String iss;
